requirements.txt
pandas
numpy
jinja2
markdown
pygments
beautifulsoup4
flask
plotly
matplotlib
bokeh
seaborn
requests
pytest
pytest-runner
pytest-cov
waitress
pyyaml
